Store world configuration in JSON
The current map-view implementation demonstrates how a JSON object
(from world-data) can store the world configuration, and we can pass
pieces of it around (such as world.regions being sent into map-view).

As part of this work, the player-view element was removed. It was a
placeholder for printing player information. At some point, we will
need some kind of player status element, showing points; for now,
however, this use case is covered by the map-view, which shows where
each player is.

One thing that leaves me nervous is that the code is fragile,
depending on the concrete structure of the JSON object. It is unclear
to me whether it should be abstracted or whether we should have unit
tests around the elements that manipulate it; either approach could
increase confidence, but only the test would prevent against
regression.