Show encounters based on world-data
There are two encounters (one in each region) that can be
completed. There is currently no checking of skills, nor are there any
computed side-effects.

As part of this work, the game-constants element was removed, with
phase constants put into app.js instead.  The element would have been
stamped out each time it was referenced, but that's a waste for a
constant. Turns out we can just get them from app with 'app.PHASES'
for example.